backend/requirements.txt
django 
djangorestframework 
channels 
channels-redis
django-allauth
axios
psycopg2-binary
awsebcli
django-cors-headers
django-rest-authtoken
djangorestframework-simplejwt
daphne
boto3 
django-storages
whitenoise
Pillow